Rename KeyState to FragmentState.
