+++ formula-android/src/main/java/com/instacart/formula/android/FragmentState.kt
@@ -1,8 +1,8 @@
-package com.instacart.formula.integration
+package com.instacart.formula.android
 
 import com.instacart.formula.fragment.FragmentKey
 
 /**
  * Defines the current render model for a specific [key].
  */
-data class KeyState(val key: FragmentKey, val renderModel: Any)
+data class FragmentState(val key: FragmentKey, val renderModel: Any)
