+++ formula-android/src/main/java/com/instacart/formula/fragment/FragmentFlowStore.kt
@@ -9,7 +9,7 @@ import com.instacart.formula.integration.Binding
 import com.instacart.formula.integration.FeatureEvent
 import com.instacart.formula.integration.DisposableScope
 import com.instacart.formula.integration.FragmentBindingBuilder
-import com.instacart.formula.integration.KeyState
+import com.instacart.formula.android.FragmentState
 import com.instacart.formula.rxjava3.toObservable
 import com.jakewharton.rxrelay3.PublishRelay
 import io.reactivex.rxjava3.core.Observable
@@ -136,7 +136,7 @@ class FragmentFlowStore(
                                 Observable.empty()
                             }
                         }.onEvent {
-                            val keyState = KeyState(fragmentId.key, it)
+                            val keyState = FragmentState(fragmentId.key, it)
                             transition(state.copy(states = state.states.plus(fragmentId to keyState)))
                         }
                     }
